Refactor Rack Attack configuration
