+++ config/initializers/rack_attack.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-require 'decidim-app/rack_attack'
+require "decidim-app/rack_attack"
+require "decidim-app/rack_attack/throttling"
+require "decidim-app/rack_attack/fail2ban"
 
 # Source: https://github.com/rack/rack-attack/issues/145#issuecomment-886180424
 class Rack::Attack
@@ -19,7 +21,10 @@ return unless Rack::Attack.enabled
 
 # Remove the original throttle from decidim-core
 # see https://github.com/decidim/decidim/blob/release/0.26-stable/decidim-core/config/initializers/rack_attack.rb#L19
-DecidimApp::RackAttack.deactivate_decidim_throttling!
+DecidimApp::RackAttack::Throttling.deactivate_decidim_throttling! do
+  Rails.logger.info("Deactivating 'requests by ip' from Decidim Core")
+  Rack::Attack.throttles.delete("requests by ip")
+end
 
 Rack::Attack.throttled_response_retry_after_header = true
 # By default use the memory store for inspecting requests
@@ -28,7 +33,7 @@ Rack::Attack.cache.store = ActiveSupport::Cache::MemoryStore.new if !ENV["MEMCAC
 
 Rack::Attack.throttled_responder = lambda do |request|
   rack_logger = Logger.new(Rails.root.join("log/rack_attack.log"))
-  throttling_limit = DecidimApp::RackAttack.throttling_limit_for(request.env["rack.attack.match_data"])
+  throttling_limit = DecidimApp::RackAttack::Throttling.throttling_limit_for(request.env["rack.attack.match_data"])
 
   request_uuid = request.env["action_dispatch.request_id"]
   params = {
@@ -41,17 +46,17 @@ Rack::Attack.throttled_responder = lambda do |request|
 
   rack_logger.warn("[#{request_uuid}] #{params}")
 
-  [429, { "Content-Type" => "text/html" }, [DecidimApp::RackAttack.html_template(throttling_limit, request.env["decidim.current_organization"]&.name)]]
+  [429, { "Content-Type" => "text/html" }, [DecidimApp::RackAttack::Throttling.html_template(throttling_limit, request.env["decidim.current_organization"]&.name)]]
 end
 
-Rack::Attack.throttle("req/ip",
-                      limit: Rails.application.secrets.dig(:decidim, :rack_attack, :throttle, :max_requests),
-                      period: Rails.application.secrets.dig(:decidim, :rack_attack, :throttle, :period)) do |req|
+Rack::Attack.throttle(DecidimApp::RackAttack::Throttling.name,
+                      limit: DecidimApp::RackAttack::Throttling.max_requests,
+                      period: DecidimApp::RackAttack::Throttling.period) do |req|
 
-  req.remote_ip unless DecidimApp::RackAttack.authorized_throttle_path?(req.path)
+  req.remote_ip unless DecidimApp::RackAttack::Throttling.authorized_path?(req.path)
 end
 
-if Rails.application.secrets.dig(:decidim, :rack_attack, :fail2ban, :enabled) == 1
+if DecidimApp::RackAttack::Fail2ban.enabled?
   # Block suspicious requests made for pentesting
   # After 1 forbidden request, block all requests from that IP for 1 hour.
   Rack::Attack.blocklist("fail2ban pentesters") do |req|
@@ -59,7 +64,7 @@ if Rails.application.secrets.dig(:decidim, :rack_attack, :fail2ban, :enabled) ==
     # so the request is blocked
     Rack::Attack::Fail2Ban.filter("pentesters-#{req.remote_ip}", maxretry: 0, findtime: 10.minutes, bantime: 1.hour) do
       # The count for the IP is incremented if the return value is truthy
-      DecidimApp::RackAttack.unauthorized_fail2ban_path?(req.path)
+      DecidimApp::RackAttack::Fail2ban.unauthorized_path?(req.path)
     end
   end
 end
